refactor write_to_buffer and use it for AArch64
